added gadm masking, updated grid code
